Commenting main app & renaming files
- Easier to know what is what
